server/evaluation/python/requirements.txt
# Python dependencies for evaluation engine
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
